added update_github_activity method to processor.py
